backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-backend"
version = "0.1.0"
description = "Phase II Todo Application Backend with FastAPI and SQLModel"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "sqlmodel==0.0.14",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
    "bcrypt==4.1.2",
    "pyjwt==2.8.0",
    "pydantic[email]==2.5.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
